Updates for types regarding millis() function

+++ src/PassedTime/PassedTime.cpp
@@ -5,15 +5,14 @@ PassedTime::PassedTime(long seconds, long millis) {
   this->_millis = millis;
 };
 
-PassedTime PassedTime::fromDistance(long startMillis, long endMillis) {
-  // result is long, because we expect non-negative arguments
-  // (they should be greater or equal zero)
-  long passedMillis;
+PassedTime PassedTime::fromDistance(unsigned long startMillis,
+                                    unsigned long endMillis) {
+  long long passedMillis;
   if (startMillis <= endMillis) {
     passedMillis = endMillis - startMillis;
   } else {
     // this can happen if millis() functions resets to zero
-    passedMillis = ((ARDUINO_MAX_MILLIS) - startMillis) + endMillis;
+    passedMillis = ((ARDUINO_MAX_MILLIS)-startMillis) + endMillis;
   }
   long passedSeconds = passedMillis / 1000;
   return PassedTime(passedSeconds, passedMillis % 1000);
